make button component
