+++ hms-app/src/components/card-details/IdeaDetails.tsx
@@ -47,6 +47,7 @@ import { getCategoryDetails } from '../../actions/CategoryActions'
 import { getSkillDetails } from '../../actions/SkillActions'
 import MoveIdeaModal from '../MoveIdeaModal'
 import IdeaCommentDetails from './IdeaCommentDetails'
+import CardButton from './CardButton'
 
 type IProps = {
   idea: Idea
@@ -136,90 +137,90 @@ export default function IdeaDetails(props: IProps) {
     </div>
   ))
 
-  const deleteSelectedIdea = () => {
-    showNotification({
-      id: 'delete-idea-load',
-      loading: true,
-      title: `Deleting "${ideaData.title}"`,
-      message: undefined,
-      autoClose: false,
-      disallowClose: false,
-    })
-    deleteIdea(instance, ideaData.id).then((response) => {
-      setDeleteModalOpened(false)
-      if (JSON.stringify(response).toString().includes('error')) {
-        setParticipantCheck(true)
-        updateNotification({
-          id: 'delete-idea-load',
-          color: 'red',
-          title: 'Failed to delete idea',
-          message: undefined,
-          icon: <X />,
-          autoClose: 2000,
-        })
-      } else {
-        setParticipantCheck(false)
-        updateNotification({
-          id: 'delete-idea-load',
-          color: 'teal',
-          title: `Deleted "${ideaData.title}"`,
-          message: undefined,
-          icon: <Check />,
-          autoClose: 2000,
-        })
-      }
-    })
-  }
+  // const deleteSelectedIdea = () => {
+  //   showNotification({
+  //     id: 'delete-idea-load',
+  //     loading: true,
+  //     title: `Deleting "${ideaData.title}"`,
+  //     message: undefined,
+  //     autoClose: false,
+  //     disallowClose: false,
+  //   })
+  //   deleteIdea(instance, ideaData.id).then((response) => {
+  //     setDeleteModalOpened(false)
+  //     if (JSON.stringify(response).toString().includes('error')) {
+  //       setParticipantCheck(true)
+  //       updateNotification({
+  //         id: 'delete-idea-load',
+  //         color: 'red',
+  //         title: 'Failed to delete idea',
+  //         message: undefined,
+  //         icon: <X />,
+  //         autoClose: 2000,
+  //       })
+  //     } else {
+  //       setParticipantCheck(false)
+  //       updateNotification({
+  //         id: 'delete-idea-load',
+  //         color: 'teal',
+  //         title: `Deleted "${ideaData.title}"`,
+  //         message: undefined,
+  //         icon: <Check />,
+  //         autoClose: 2000,
+  //       })
+  //     }
+  //   })
+  // }
 
-  const closeEditModal = (isOpened: boolean) => {
-    setEditModalOpened(isOpened)
-  }
+  // const closeEditModal = (isOpened: boolean) => {
+  //   setEditModalOpened(isOpened)
+  // }
 
-  const deleteModal = (
-    <Modal
-      centered
-      opened={deleteModalOpened}
-      onClose={() => setDeleteModalOpened(false)}
-      withCloseButton={false}
-    >
-      <Text className={classes.text}>
-        Are you sure you want to delete this idea?
-      </Text>
-      <Text className={classes.title}>Title: {ideaData.title}</Text>
-      <Button
-        style={{ backgroundColor: DELETE_BUTTON_COLOR }}
-        onClick={() => deleteSelectedIdea()}
-      >
-        Yes, delete this idea
-      </Button>
-      <Text className={classes.text}>
-        (This window will automatically close as soon as the idea is deleted)
-      </Text>
-    </Modal>
-  )
+  // const deleteModal = (
+  //   <Modal
+  //     centered
+  //     opened={deleteModalOpened}
+  //     onClose={() => setDeleteModalOpened(false)}
+  //     withCloseButton={false}
+  //   >
+  //     <Text className={classes.text}>
+  //       Are you sure you want to delete this idea?
+  //     </Text>
+  //     <Text className={classes.title}>Title: {ideaData.title}</Text>
+  //     <Button
+  //       style={{ backgroundColor: DELETE_BUTTON_COLOR }}
+  //       onClick={() => deleteSelectedIdea()}
+  //     >
+  //       Yes, delete this idea
+  //     </Button>
+  //     <Text className={classes.text}>
+  //       (This window will automatically close as soon as the idea is deleted)
+  //     </Text>
+  //   </Modal>
+  // )
 
-  const editModal = (
-    <Modal
-      centered
-      opened={editModalOpened}
-      onClose={() => setEditModalOpened(false)}
-      withCloseButton={false}
-      size='55%'
-    >
-      <Text className={classes.title}>Edit Idea</Text>
-      <IdeaForm
-        ideaId={ideaData.id}
-        idea={ideaData}
-        context={IdeaFormType.Edit}
-        ownerId={ideaData.owner ? ideaData.owner.id : ''}
-        hackathon={ideaData.hackathon!}
-        setOpened={closeEditModal}
-      />
-      <Text className={classes.text}>
-        (This window will automatically close as soon as the idea is changed)
-      </Text>
-    </Modal>
-  )
+  // const editModal = (
+  //   <Modal
+  //     centered
+  //     opened={editModalOpened}
+  //     onClose={() => setEditModalOpened(false)}
+  //     withCloseButton={false}
+  //     size='55%'
+  //   >
+  //     <Text className={classes.title}>Edit Idea</Text>
+  //     <IdeaForm
+  //       ideaId={ideaData.id}
+  //       idea={ideaData}
+  //       context={IdeaFormType.Edit}
+  //       ownerId={ideaData.owner ? ideaData.owner.id : ''}
+  //       hackathon={ideaData.hackathon!}
+  //       setOpened={closeEditModal}
+  //     />
+  //     <Text className={classes.text}>
+  //       (This window will automatically close as soon as the idea is changed)
+  //     </Text>
+  //   </Modal>
+  // )
 
   const addParticipant = async (
     action = createIdeaParticipant,
@@ -468,6 +469,7 @@ export default function IdeaDetails(props: IProps) {
                   <Text className={classes.label}>Description</Text>
                   <Text className={classes.text}>{ideaData.description}</Text>
                 </Card.Section>
+
                 <Card.Section className={classes.borderSection}>
                   <Text className={classes.label}>Problem</Text>
                   <Text className={classes.text}>{ideaData.problem}</Text>
@@ -477,33 +479,19 @@ export default function IdeaDetails(props: IProps) {
                   <Text className={classes.label}>Goal</Text>
                   <Text className={classes.text}>{ideaData.goal}</Text>
                 </Card.Section>
+
+                {/* <Card.Section className={classes.borderSection}>
+                  <Text className={classes.label}>Skills</Text>
+                  <Text className={classes.text}>{getSkillDescription(ideaData.id)}</Text>
+                </Card.Section> */}
               </div>
 
               {type === IdeaCardType.Admin ||
-                type === IdeaCardType.Owner ||
+                type === IdeaCardType.Owner &&
                 (ideaData.owner?.id === user?.id && (
-                  <Group position='left' mt='xl'>
-                    {deleteModal}
-                    <Button
-                      style={{
-                        backgroundColor: DELETE_BUTTON_COLOR,
-                      }}
-                      onClick={() => setDeleteModalOpened(true)}
-                    >
-                      Delete
-                    </Button>
-                    {editModal}
-                    <Button
-                      style={{
-                        backgroundColor: JOIN_BUTTON_COLOR,
-                      }}
-                      onClick={() => setEditModalOpened(true)}
-                    >
-                      Edit
-                    </Button>
-                    <FinalVideoUploadModal idea={ideaData} />
-                    <MoveIdeaModal idea={ideaData} />
-                  </Group>
+                  <CardButton
+                    idea={props.idea}
+                  />
                 ))}
             </Accordion.Panel>
           </Accordion.Item>
@@ -621,7 +609,7 @@ export default function IdeaDetails(props: IProps) {
                   <Accordion.Panel>
                     <div>{ideaDetails()}</div>
 
-                    {type === IdeaCardType.AllIdeas && (
+                    {type === IdeaCardType.AllIdeas && (ideaData.owner?.id === user?.id && (
                       <Group
                         mt='xs'
                         position={'right'}
@@ -656,35 +644,20 @@ export default function IdeaDetails(props: IProps) {
                           >
                             {voteCheck ? 'Remove Vote' : 'Vote for Idea'}
                           </Button>
+                          
                         )}
+                        <CardButton
+                          idea={props.idea}
+                        />
                       </Group>
-                    )}
+                    ))}
 
                     {type === IdeaCardType.Admin ||
                       type === IdeaCardType.Owner && 
                       (ideaData.owner?.id === user?.id && (
-                        <Group position='left' mt='xl'>
-                          {deleteModal}
-                          <Button
-                            style={{
-                              backgroundColor: DELETE_BUTTON_COLOR,
-                            }}
-                            onClick={() => setDeleteModalOpened(true)}
-                          >
-                            Delete
-                          </Button>
-                          {editModal}
-                          <Button
-                            style={{
-                              backgroundColor: JOIN_BUTTON_COLOR,
-                            }}
-                            onClick={() => setEditModalOpened(true)}
-                          >
-                            Edit
-                          </Button>
-                          <FinalVideoUploadModal idea={ideaData} />
-                          <MoveIdeaModal idea={ideaData} />
-                        </Group>
+                        <CardButton
+                          idea={props.idea}
+                        />
                       ))}
                   </Accordion.Panel>
                 </Accordion.Item>
